Renamed ConsensusRules folder to BasePoAFeatureConsensusRules
