+++ src/Stratis.Bitcoin.Features.PoA.Tests/Rules/PoAHeaderDifficultyRuleTests.cs
@@ -6,7 +6,7 @@ using NBitcoin;
 using NBitcoin.BouncyCastle.Math;
 using Stratis.Bitcoin.Consensus;
 using Stratis.Bitcoin.Consensus.Rules;
-using Stratis.Bitcoin.Features.PoA.ConsensusRules;
+using Stratis.Bitcoin.Features.PoA.BasePoAFeatureConsensusRules;
 using Stratis.Bitcoin.Tests.Common;
 using Xunit;
