+++ src/Stratis.Bitcoin.Features.PoA.Tests/Rules/PoAIntegritySignatureRuleTests.cs
@@ -5,7 +5,7 @@ using NBitcoin;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Consensus;
 using Stratis.Bitcoin.Consensus.Rules;
-using Stratis.Bitcoin.Features.PoA.ConsensusRules;
+using Stratis.Bitcoin.Features.PoA.BasePoAFeatureConsensusRules;
 using Xunit;
 
 namespace Stratis.Bitcoin.Features.PoA.Tests.Rules
